requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM Providers
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
langchain-huggingface>=0.0.1
langchain-community>=0.2.0

# Embeddings and Vector Stores
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Document Processing
PyPDF2>=3.0.1

# Web Search
duckduckgo-search>=3.9.0
google-api-python-client>=2.100.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0

# Development (optional)
black>=23.12.0
flake8>=6.1.0
pytest>=7.4.0

# Additional useful packages
tqdm>=4.67.0
pydantic>=2.5.0
